First test passing.
